ml_service/budget_estimation/requirements.txt
# Budget Estimation ML Service Dependencies

# Core ML libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: Advanced ML (uncomment if needed)
# tensorflow==2.14.0
# torch==2.1.0
# xgboost==2.0.1
# lightgbm==4.1.0
